[Fixed] Bug:0000083: Low Memory Mode will prevent Transaction Clerk, Foundation Manager, Balance Indexer from running during Foundation repair

+++ server/v4.22/foundation.php
@@ -325,6 +325,7 @@ if(($next_generation_cycle - time()) > 60 && (time() - $current_generation_cycle
 							mysqli_query($db_connect, "UPDATE `main_loop_status` SET `field_data` = '1' WHERE `main_loop_status`.`field_name` = 'block_check_back' LIMIT 1");
 							mysqli_query($db_connect, "UPDATE `main_loop_status` SET `field_data` = '$foundation_time_start' WHERE `main_loop_status`.`field_name` = 'foundation_block_check_start' LIMIT 1");
 							mysqli_query($db_connect, "UPDATE `main_loop_status` SET `field_data` = '$foundation_time_end' WHERE `main_loop_status`.`field_name` = 'foundation_block_check_end' LIMIT 1");
+							$foundation_task = 1;
 						}
 					}
 				}
@@ -518,7 +519,7 @@ else
 	// Memory Management Check
 	$low_memory_mode = mysql_result(mysqli_query($db_connect, "SELECT field_data FROM `main_loop_status` WHERE `field_name` = 'low_memory_mode' LIMIT 1"));
 
-	if($low_memory_mode == 1)
+	if($low_memory_mode == 1 && $foundation_task == 0)
 	{
 		// Exit to release any RAM being held, the Main Program will restart this script afterwards
 		mysqli_query($db_connect, "DELETE FROM `main_loop_status` WHERE `main_loop_status`.`field_name` = 'foundation_heartbeat_active'");
